js: package.json: rename JavaScript package to '@intc/stats-collect'
The package name 'stats-collect-components' was unscoped and could be
reserved on the public 'npm' registry by an external actor. This
resulted in a dependency-confusion vulnerability.

To remove this vulnerability, this patch renames the package to include
the '@intc' organisation scope. This means that the package can't be
populated on any registry by someone outside of the Intel organisation.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>