_DTabBuilder: prevent 'stats-collect report' crashing
Before this patch, if the reference result did not contain data for the
metric in the plot being checked by '_skip_metric_plot()', then
'stats-collect' would crash.

This patch fixes that behaviour by taking the sample datapoint from the
first result with data for the metric rather than just the first result.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>