statscollectlibs: add labels support
Labels are a mechanism for synchronizing workload stages and metrics with
statistics data. Statistics are collected independently of the workload, but
sometimes it is necessary to match a piece of collected statistics with a
workload stage. For example, if the workload has a warmup stage, one can add a
"skip" label before the worm up, and statistics visualization tools will
exclude it from visualization.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>