_ACPowerTabBuilder: fix bad CSV causing 'stats-collect report' crash
If there is a bad CSV AC Power statistic file, we do not want
'stats-collect report' to crash. It should just fail on that tab and
continue trying to generate other tabs.

Recently, the 'dtype' parameter was added to the 'read_csv()' call. This
means that columns are converted to 'dtype' before being parsed into the
dataframe. If that conversion fails, it produces a 'ValueError', not the
usual parsing 'pandas.errors.ParserError'.

If a 'ValueError' was raised, it was not being converted to a
'pepclibs.helperlibs.Exceptions.Error' and therefore avoided the
'stats-collect' error handling and caused a crash in 'stats-collect
report'. This patch adds 'ValueError' to the suitable 'try-except'
statement so that those errors can be properly handled.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>