_StatsCollectStart: move 'stats_res' creation to '_SpecStatsCollect'
Before this patch, all tools needed to create a statistics result
themselves and then pass it to 'StatsCollect'.

This patch reduces the code duplication by moving the statistics result
creation to 'SpecStatsCollect'. This means that after this patch, tools
just need to create an instance of 'SpecStatsCollect' and it will handle
the statistics result creation.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>