tests: common: fix pytests crashing
In a previous patch, the 'TOOLNAME' constant was moved
to 'ToolInfo' from '_StatsCollect'. The 'TestRunner' class has been
changed so that the caller can still provide the 'TOOLNAME' in a new
parameter.

This means that the pytest helper 'run_stats_collect()' crashes as
before this patch, it did not provide the 'TOOLNAME' constant in any
way.

This patch updates the 'TestRunner' caller in 'common.py' to use the new
parameter and fixes the pytests.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>