_STCAgent: fix 'add_label()' crashing if no statistics are enabled
If no statistics are enabled, it means that 'STCAgent' has not been
connected to any 'stc-agent'. Before this patch, 'add_label()' would
always try to send a command to 'stc-agent', even if the class was not
connected. This caused an error and would crash the 'stats-collect'
tool.

This patch prevents that error by changing 'add_label()' to check if
'STCAgent' has any enabled statistics before sending the command to add
a label.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>