_SpecStatsCollect: make 'info.yml' labels path to be relative to outdir
The read-only result classes which read the 'info.yml' file are not
aware of the 'agent.statsdir' directories. This makes the labels path
useless to them.

For this reason, this patch changes the path dumped to 'info.yml' to be
relative to 'outdir' so that those read-only result classes can open the
labels files based on the path they read in the 'info.yml' file.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>