_DeployPyHelpers: search for deployables first in process' directory
This patch fixes the following problem.

Developer modifies 'stc-agent' in the git tree, and then runs 'stats-collect
deploy -H abc'. The expectation is that the modified version of 'stc-agent'
will be deployed to host 'abc'. However, this is not always the case. If there
is a locally installed 'stc-agent' available in 'PATH', say
'/usr/bin/stc-agent', then this agent will be deployed. As a result, the
developer will not be able to test the 'stc-agent' modification.

Fix the problem by first looking for the 'stc-agent' in the directory
'stats-collect' runs from. If it runs from a git repository, then 'stc-agent'
from the same git repository will be used.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>