js: file-preview: fix 'sysinfo' tabs not loading locally
This patch fixes a bug where 'sysinfo' tabs, specifically
'file-previews' would not load in reports viewed locally.

The patch fixes the bug by converting 'file-previews' to also use 'File'
objects which are resolved by the 'report-page' component (similarly to
summary table elements). The files are resolved using 'fetch' if the
report is viewed remotely or from uploaded files if the report is viewed
locally.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>