_MCPUL2TabBuilder: fix tab generation crash if measured CPU is not 0
Package-level metrics are only included in the first row of a package in
a turbostat table. This means that if another CPU is measured, the
values for those metrics are not visible. This patch adjusts this by
making the package values available no matter which cpu is measured.

This fixes the tab generation crashing when the measured CPU is not the
first in a package.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>