_StatsCollect: make 'stats-collect start --cpunum' default to 'None'
The 'cpunum' option is not commonly used and the default value of 0
might be confusing to users who did not specify a value for the option.
Specicically, the "Measured CPU" turbostat tab in HTML reports can be
confusing if the user did not provide a "Measured CPU".

Therefore, this patch changes the default value to 'None' so that no
"Measured CPU" tab will be generated unless a user specifies one.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>