_TurbostatL2TabBuilderBase: add 'C-states' 'Requested' 'Count' tab
This patch adds a new data tab containing data on the number of times
various C-states were requested.

Note, it splits the 'C-states' 'Requested' tab into two data tabs: one
for residency data (which existed before this patch) and one for the
count of C-state requests (added in this patch).

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>