_MCPUL2TabBuilder: cleanup
It is cleaner and simpler to use one variable rather than using one
variable 'cpunum' to determine if the result has a specific tested CPU,
and another variable 'mcpu' to give the CPU number to the
'pandas.DataFrame' builder.

This patch makes the 'MCPUl2TabBuilder.__init__()' method easier to read
and understand.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>