StatsCollect: simplify constructor arguments
The 'StatsCollect' and related classes take 4 additional argument for the sake
of creating a 'WORawResult()' object.

Simplify it so that the caller creates 'WORawResult' object and passes it down
to 'StatsCollect' constructor. This will scale better - if we add more options
to 'WORawResult', we won't need to add yet another argument to multiple
'StatsCollect' family classes.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>