tests: common: set 'STATS_COLLECT_DATA_PATH' for tests
This patch fixes an issue where running pytest caused failures if there
were any changes to data files. It does this by setting the
'STATS_COLLECT_DATA_PATH' environment variable to the project root of
the tested installation.

This means that it will always use the data files present in the
instance of 'stats-collect' being tested.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>