tests: common: remove Python path hack
This patch removes a hack which was used by tests to make sure that the
the local '_StatsCollect' module was used over the '_StatsCollect'
module in the 'wult' project which may have also been on the Python
path.

The hack was written when 'stats-collect' code was also present in
the 'wult' project. 'stats-collect' has now been completely split from
the 'wult' project, so this hack is no longer necessary.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>