_DTabBuilder: turn skipped plot messages into debug messages
These messages are often printed a lot during report generation. The
report visually also tells users if a plot was skipped because every
datapoint in all metrics was the same.

This makes the info messages redundant, so this patch changes them to
debug messages as it might be useful for developers to quickly see if
any reports were skipped without opening and loading the report.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>