package.json: remove 'babel' dependency
'babel' depended on 'semver' version '<7'. That version had a moderate
security alert (CVE-2022-25883). There is currently no version of
'babel' which avoids this vulnerability.

'babel' was an extraneous dependency for 'stats-collect', only used for
linting source code during development. This patch simplifies the
'eslint' configuration to use the default parser for linting (and not
babel). This means that the development configuration is simpler and
safer by removing the 'babel' dependency.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>