Runner: fix '--time-limt' handling
When a command runs with '--time-limit' and does not exit after the specified
time limit, it gets killed. After this, 'stats-collect' is supposed to succeed.
Howerver, it fails with a misterious error message like this:

stats-collect: error: there was an error running command 'sleep 10000000':

Fix the issue by breaking out for the loop after the command is killed.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>