_Common: separate report generation logic to '_StatsCollectReport'
'stats-collect' HTML reports expand on the base 'HTMLReport' class to
specify the contents of all 'stats-collect' reports. With time, the
logic used to specify the contents of those reports has grown.

To make it easier to maintain, this patch separates all of the
'stats-collect' HTML report functionality into its own class
'_StatsCollectReport'.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>