FilePreviewBuilder: remove extra sub-directories from 'diffs' directory
This patch changes 'FilePreviewBuilder' so that rather than preserving
the directory structure of the raw files when placing '.diff' files
inside the report directory, the class will now just place all '.diff'
files directly in the 'diffs' directory.

This simplifies the structure and contributes to making the entire
report directory easier to navigate.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>