HTMLReport: use a variable for the "SysInfo" tab name
Rather than hard-coding the tab name in multiple places of
'_generate_sysinfo_tab()', it is easier to change in future if the name
is stored in a variable.

Therefore this patch defines the tab name as "SysInfo" and replaces all
instances of hard-coding the name with references to the variable.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>