_StatsTabBuilder: add 'name' class property
This patch adds the 'name' class property to the 'StatsTabBuilder' class
so that the name does not need to be hard-coded in multiple places.
Instead, the class property can be referenced. This will also make it
easier to change in future if necessary.

Furthermore, this change is also consistent with other tab builder
classes which also have a 'name' class property.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>