_TabBuilderBase: '_build_ctab()': build tabs if any results have data
This patch changes '_build_ctab()' so that the method generates data tabs for
metrics in the 'tab_hierarchy' which only one or more results have data.

Specifically this is in preparation to change the way 'turbostat' tabs
are generated so that 'turbostat' tabs will be generated for metrics
present in any result rather than just metrics present in all results.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>