htmlreport: remove duplicate parameter docstring definitions
The 'basedir' and 'outdir' parameters are defined in the docstring of
many different classes. To reduce duplication, this patch replaces all
but one of them with a reference to the remaining definition. This means
that it is only defined in one place and can be easily changed in future
if necessary.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>