README.md: add a section on basic usage
Since there is no web documentation for 'stats-collect', it would be
helpful for people who visit the 'stats-collect' repository to have some
idea of the basic 'stats-collect' workflow. Therefore this patch adds a
'Basic Usage' section which explains how to use 'stats-collect' and
points to the man page for more detailed information.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>