Runner: kill 'cmd' if it is still running when 'close()' is called
This patch changes 'Runner.close()' so that it tries to kill the process
running 'cmd' if it is still executing when 'close()' is called. For
example if the user causes a 'KeyboardInterrupt' before the command has
finished executing, then try to kill the process running 'cmd'. This
should prevent orphaning the process and leaving them runnning on the
SUT.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>