Runner: fix 'stats-collect start --stats=none' crashing
This patch fixes a regression which caused
'stats-collect start --stats=none' to crash. The crash was caused by the
introduction of a minimum run duration based on the maximum statistics
collection interval.

The regression was introduced in:
 - commit 58a52cfbaf54fe78d5baa5559549bdc3683ce48e:
    Runner: raise error if 'cmd' finishes before collecting enough data

When '--stats=none' is used, there are no statistics collectors for
which the interval should be considered and hence, there should not be a
minumum run duration.

This patch changes the 'Runner' class to only enforce the minimum run
duration if the statistics collector was instantiated.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>