_Plot: change '_is_numeric_col()' to also detect time-series data
The helper function '_is_numeric_col()' is used to dictate the format of
plots. Specifically it is used to determine if the data is categorical
or scalar. This patch changes the function to also return 'True' if the
data is time-series data as this should be presented in the same way as
other numerical data.

This is in preparation to improve the way that 'Time Elapsed' on
diagrams is shown in hover-text and on axes.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>