dfbuilders: change "Time Elapsed" representation from 'ks' to 'hh:mm:ss'
This patch changes the way that the "Time Elapsed" metric in "AC Power",
"IPMI" and "Turbostat" tabs is represented on plot axes and in plot
hovertext. Rather than showing "ks", the tabs will now show the time
elapsed in 'hh:mm:ss' format. This makes the time much more intuitive
for researchers using the reports.

The change is made by converting the "Time Elapsed" data to use the
'numpy' 'datetime64' format.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>