htmlreport: reduce the file size of diagrams with hover text
This patch changes the way that plots are generated using 'plotly'. This
patch changes the 'Histogram' and 'ScatterPlot' classes to no longer
pass entire statistics 'pandas.DataFrames' as 'customdata' to 'plotly'.
This is because 'plotly' then includes the entire 'pandas.DataFrame' in
the resultant plot HTML file.

'customdata' was originally used to populate the hover text. By
formatting hover text for values from the 'pandas.DataFrame' manually,
'plotly' only dumps the required hover text rather than the text and
'pandas.DataFrames'.

The result of this change is that the file size of diagrams with hover
text are significantly smaller.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>