_SpecStatsCollect: fix 'ipmi' abstract stname resolution failing
This patch fixes a functional error where 'SpecStatsCollect' would log
that it was disabling 'ipmi-inband' in favour of 'ipmi-oob', however it
then proceeded to collect both in-band and out-of-band 'ipmi' data.

The patch fixes it by moving the 'configure' step after the 'ipmi'
abstract names have been resolved. This means that when 'set-stats' is
called during the 'configure' stage, it will be called with only one of
the two 'ipmi' stnames.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>