_TotalsL2TabBuilder: fix totals hardware tabs not getting added to tab
This patch fixes a regression in a recent rework which caused the
turbostat totals tabs to not be added to the tab configuration.

The bug happened because the class was checking if a whole
'PackageCSDef' instance was in 'common_metrics', when 'common_metrics'
is populated with metric names. Therefore, this patch fixes the bug by
checking if the metric name of each 'PackageCSDef' is in common metrics.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>