_DTabBuilder: replace 'metric_def' constructor parameter with 'tabname'
This patch replaces the 'metric_def' parameter in
'DTabBuilder.__init__()' with the 'tabname' parameter. This is because
the 'metric_def' is only used to dictate the name of the tab.

Therefore, changing it directly to 'tabname' makes this clearer and also
makes it more intuitive to customise the title of the tab.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>