TabConfig: replace the 'tab_mdef' parameter with 'name' parameter
It is a much simpler concept and more intuitive to require a name for
the tab than a 'tab_mdef'. Therefore, this patch replaces the 'tab_mdef'
config attribute with 'name'. This is more consistent with 'DTabBuilder'
and will make it easier to customise tabs in future when users can specify
custom tabs.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>