_Plot: improve unit handling on plot axes
This patch changes the plot classes so that data points with units
containing SI-unit prefixes, will automatically be scaled to use the
most human-readable SI-unit prefix.

For example, if a data point for the 'turbostat' metric 'Busy_MHz' has a
value '1200 MHz', then it will be scaled so that the plot axes show the
data point as '1.200 GHz'.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>