_DFBuilderBase: fix a bug where loading labels causes lines on plots
This patch changes the way that '_DFBuilderBase' loads labels into
statistics 'pandas.DataFrames'. It removes an unnecessary step which set
all datapoints to default to 'None' for labels.

By setting values to 'None', this changed the 'pandas.DataFrame.dtype'
to 'Object'. In turn, this meant that plots thought the data was not
numerical, and would attempt to display it using line markers rather
than dots. This patch fixes this behaviour by simply removing the
extraneous step.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>