_TurbostatL2TabBuilderBase: remove references to 'common_metrics'
Previously 'turbostat' tabs were adjusted to include all metrics, not
just those common to every result. However, references to the list of
metrics to be included in the results were still labelled as
'common_metrics' in code.

This patch changes these references so that they no longer indicate that
the metrics are limited to those which are common to every result.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>