ipmi-helper: replace human-readable timestamp with epoch timestamp
The human-readable timestamp is time-zone adjusted. This makes it
impossible to apply labels which are saved with a time-zone agnostic
timestamp. Additionally, the format is also not consistent with other
statistics files.

This patch changes the format so that it is with other stats files and
to allow labels to be applied in HTML reports to 'ipmi' data.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>