WORawResult: change path deletion message to debug message
The message is not useful to users, only to developers. If the user
used the '--stats' option to specify which statistics should be
collected (if any), then they will be informed through an error or
warning by the tool earlier than this message.

Additionally, the details of a specific directory being deleted is not
relevant to the user as they were not aware the directory existed
beforehand, only which statistics they have requested.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>