TurbostatDFBuilder: remove 'reduced_tstat'
In several places, 'TurbostatDFBuilder' classes reduce the metrics which
are included in 'turbostat pandas.DataFrames' to only include metrics
which also appear in the 'totals' dictionary (formed by the parser to
contain metrics from the turbostat summary output).

There is no reason for this restriction as it is also redundant because
all metrics for CPUs are also included in the summary.

Therefore the point of this patch is to stop restricting the metrics
which can be loaded into a turbostat 'pandas.DataFrame'.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>