_TurbostatL2TabBuilderBase: rename '_get_ctab_cfg()'
Rename '_get_ctab_cfg()' to '_get_default_tab_cfg()'. This name is more
appropriate as it is consistent with other helper functions in that it
appends an underscore to the function that it 'helps'.

The '_get_ctab_cfg()' method name will also be more useful in future to
use for a more generic method.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>