TurbostatParser: include 'nontable' in all parsing iterations
Inclue the 'nontable' dictionary in all iterations of the turbostat
parser rather than only the first. This is more intuitive than the
'nontable' dictionary being empty in every iteration after the first.

This will facilitate the addition of a sanity check to use TDP to
validate the values such as 'PkgWatt' and 'RAMWatt'.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>