statscollect: add workload data reference to the intro table
Complex workloads, such as SPECJbb generate own data and own reports. Make it
possible to add references to that workload data from the intro table.

Add another "Workload data" row to the table, just after the "Logs" row, and
put a ling to that row for every raw result.

The workload wada is supposed to reside in the raw result directory under the
"wldata" sub-directory. Create a intro table reference only When this
sub-directory exists.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>