RORawResult: remove old raw results support with no stats path
In very old versions of stats-collect the statistics file paths were not stored
in 'info.yml'. For backward compatibility, they are also hard-codes in
'load_stat()' method calls.

Remove the hard-coded names in order to simplify the code a bit. This will
break compatibility with very old raw results though.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>