statscollectlibs: crop SPECjbb2015 results by absolute time-stamp
Trimming by the relative time-stamp is not reliable, and gives unsatisfactory
results and lines shifted relative to each other when the initial configuration
time (before SPECjbb ven starts) is very different between the results.

Use absolute time-stamp instead, which fixes the issue. But absolute
time-stamps are only available in the controller log file. On the other hand,
controller log file does not include critical-jOPS. Because of this, parse both
the controller log and stdout files.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>