stats-collect: fix log handling on localhost
When running stats-collect locally, and collecting
data fails, logs are incorrectly removed. The logs
may include hints about failure, and should not be
removed.

Fix is to save data to temporary directory first,
and in the end, copy data to final output
directory. In case of error, temporary directory
is not removed, and is available for later
inspection.

Signed-off-by: Antti Laakso <antti.laakso@intel.com>