_SysInfoDTabBUilderBase: skip individual file preview
Fix a regression that is caused by adding the 'pepc pmqos' file preview to the
SysInfo tab - when visualizing the old results, the 'pepc pmqos' raw data files
are missing, and the entire 'pepc' tab is skipped as a result.

Fix this by skipping just the 'pepc pmqos' file preview instead, keep the
'pepc' tab with other file previews, such as 'pepc cstates'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>