FilePreviewBuilder: skip only too big files, not entire file previews
Problem 1: When a file preview file is too big, the entire file preview is
skipped.
Problem 2: The file sise is checked only if it not in the output directory. But
file size check is about limiting the final HTML file preview element size, and
it does not matter where the file resides.

Fix both problems.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>