stats-collect start: remove --cpus option
This option does not make much sense, it is just a recommendation on what to
visualize. But what to visualize is not about 'stats-collect start', it is
about 'stats-collect report'.

Remove the option. In the future, it can be returned with different semantics:
collect statistics only for limited amount of CPUs, instead of collecting for
all CPUs.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>